Add an error handler module to rescue frequently raised exceptions
Our application frequently raises exceptions that are "expected". We are aware of those errors and decided that they are safe to ignore.

Because our monitoring system sends us notifications for each application error, we run the risk to miss important ones if we are flooded with errors we don't need to worry about.

This commit adds an `ErrorHandler` module that can be mixed-in in any class that is known to frequently raise expected errors. The module provides a method to wrap a block of code, and, conditionally rescue exceptions raised within that block.

Use `handle_errors` to wrap blocks of code that are known to raise expected exceptions. Provide an array of rules as `Procs` to indicate which exceptions should be rescued. If any rule matches (OR), the exception will be rescued. Otherwise, it will be re-raised. If the same rules are used for multiple classes, they can be extracted into a module.

```ruby
class JsonParser
  include ErrorHandler

  def parse
    handle_errors([->(e) {e.is_a?(IOError)}]) do
      # Code that often raises a IOError
    end
  end
end
```